Refactored to remove name dummy
