+++ tests/conftest.py
@@ -20,5 +20,5 @@ def app() -> Flask:
 
 @pytest.fixture(scope="session")
 def load_test_data():
-    fund_data.FUNDS_DUMMY_DAO.load_dummy(copy.deepcopy(TEST_FUND_DATA))
-    round_data.ROUNDS_DUMMY_DAO.load_dummy(copy.deepcopy(TEST_ROUND_DATA))
+    fund_data.FUNDS_DAO.load_data(copy.deepcopy(TEST_FUND_DATA))
+    round_data.ROUNDS_DAO.load_data(copy.deepcopy(TEST_ROUND_DATA))
